Move LocalFile to testing
- LocalFile should be for internal testing only
- Remove gridds for downloading grid datasets
- update docs
