+++ README.md
@@ -68,7 +68,7 @@ The `xAOD` code only renders the `func_adl` expression as a ROOT file. The ROOT 
 - If a `Select` sequence of a `tuple` is the last `func_adl` expression, then a file called `xaod_output.root` will be generated, and it will contain a `TTree` called `xaod_tree` with a columns named `col1`, `col2`, etc.
 - If a `Select` sequence of dictionary's is the last `func_adl` expression, then a file called `xaod_output.root` will be generated, and it will contain a `TTree` called `xaod_tree`, with column names taken from the dictionary keys.
 
-`ServiceX` (and the [`servicex` frontend package](https://pypi.org/project/servicex/)) can convert from ROOT to other formats like a `pandas.DataFrame` or an `awkward` array. This package contains a `LocalFile` object which can do something similar, but it is built only for testing, and only for ROOT file output.
+`ServiceX` (and the [`servicex` frontend package](https://pypi.org/project/servicex/)) can convert from ROOT to other formats like a `pandas.DataFrame` or an `awkward` array.
 
 ## Testing and Development
